changing StackHeaderToolbarMenuItemOptionsAndroid to StackHeaderToolbarMenuElementOptionsAndroid, and renaming scenario-descriptions files to scenario-description to follow naming pattern
