+++ apps/src/tests/single-feature-tests/stack-v5/test-stack-toolbar-menu-commands-android/scenario-description.ts
