+++ apps/src/tests/single-feature-tests/stack-v5/test-stack-toolbar-menu-disabled-android/scenario-description.ts
