+++ apps/src/tests/single-feature-tests/stack-v5/test-stack-toolbar-menu-icon-android/scenario-description.ts
